# docs/swagger.yaml
basePath: /
definitions:
...
  models.CardItemFull:
    properties:
      code:
        type: string
      extension:
        type: string
      formula:
        type: string
      id:
        type: integer
      imagePath:
        type: string
      index:
        type: integer
      shared:
        type: boolean
      text:
        type: string
      type:
        $ref: '#/definitions/schema.CardItemType'
      userId:
        type: integer
      width:
        type: string
    type: object
...
  models.CardsByQueryRequest:
    properties:
      group:
        $ref: '#/definitions/schema.CardsGroup'
      id:
        type: integer
      priority:
        $ref: '#/definitions/schema.CardsPriority'
      query:
        type: string
    required:
    - id
    type: object
...
  models.MoveSharedNodeToUserRequest:
    properties:
      deep:
        type: boolean
      userId:
        type: integer
    required:
    - userId
    type: object
...
  models.RemoveSharedNodeFromUserRequest:
    properties:
      deep:
        type: boolean
      userId:
        type: integer
    required:
    - userId
    type: object
...
host: localhost:3033
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Memory nodes, cards, and card items API server.
    **Authentication:** Click **Authorize**, enter your email as username and your password, then Authorize. Swagger will call `POST /users/token` and attach the access token as a Bearer header.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: MemoryGuard API
  version: "1.0"
paths:
...
  /admin/memory-node/{id}/move-to-user:
    post:
      consumes:
      - application/json
      description: Admin-only. Grants a shared memory node and its shared cards to
        a user. When deep=true, also grants shared descendants recursively.
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      - description: Target user and deep flag
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MoveSharedNodeToUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Move shared memory node to user (admin)
      tags:
      - memory-nodes
  /admin/memory-node/{id}/remove-from-user:
    post:
      consumes:
      - application/json
      description: Admin-only. Revokes a user's access by deleting memory_node_users
        / card_users (and card_user_counts) links. Does not delete the node or cards.
        When deep=true, also revokes shared descendants recursively.
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      - description: Target user and deep flag
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RemoveSharedNodeFromUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Remove shared memory node from user (admin)
      tags:
      - memory-nodes
  /card-item/{id}:
    delete:
      description: Deletes a card item by ID owned by the authenticated user
      parameters:
      - description: Card item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Delete card item
      tags:
      - card-items
    get:
      description: Returns a card item by its ID for the authenticated user
      parameters:
      - description: Card item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardItemFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get card item by ID
      tags:
      - card-items
  /card-items:
    get:
      description: Returns all card items for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardItemFull'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: List card items
      tags:
      - card-items
  /card/{id}:
    get:
      description: Returns a card with resolved question/answer card items for the
        authenticated user
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get card by ID
      tags:
      - cards
  /cards:
    get:
      description: Returns cards for the given JSON ids query parameter owned by the
        authenticated user
      parameters:
      - description: JSON array of IDs, e.g. [1,2,3]
        in: query
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: List cards by IDs (query)
      tags:
      - cards
  /cards-by-query:
    post:
      consumes:
      - application/json
      description: Returns cards for a memory node (or priority/group), filtered by
        a simple query string
      parameters:
      - description: Query request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CardsByQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Filter cards by query
      tags:
      - cards
  /decrease-card-count/{id}:
    put:
      description: Decrements the card count by 1 (not below 0)
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Decrease card count
      tags:
      - cards
  /delete-cards:
    post:
      consumes:
      - application/json
      description: Deletes owned cards by IDs (including shared cards owned by the
        user). For shared cards the user only has a grant to, removes the grant/count
        records instead of deleting the card.
      parameters:
      - description: List of card IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Delete cards
      tags:
      - cards
  /get-card-items:
    post:
      consumes:
      - application/json
      description: Returns multiple card items by their IDs for the authenticated
        user
      parameters:
      - description: List of card item IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardItemFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get card items by IDs
      tags:
      - card-items
  /get-cards:
    post:
      consumes:
      - application/json
      description: Returns multiple cards by their IDs for the authenticated user
      parameters:
      - description: List of card IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get cards by IDs
      tags:
      - cards
  /get-memory-nodes:
    post:
      consumes:
      - application/json
      description: Returns multiple memory nodes by their IDs for the authenticated
        user
      parameters:
      - description: List of memory node IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MemoryNodeFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get memory nodes by IDs
      tags:
      - memory-nodes
  /increase-card-count/{id}:
    put:
      description: Increments the card count by 1
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Increase card count
      tags:
      - cards
  /memory-node-by-alias/{alias}:
    get:
      description: Finds a memory node by one of its aliases for the authenticated
        user
      parameters:
      - description: Alias
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MemoryNodeFull'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get memory node by alias
      tags:
      - memory-nodes
  /memory-node/{id}:
    delete:
      description: Deletes a memory node owned by the authenticated user. If the node
        is shared, also deletes owned child nodes, their cards, and all memory_node_users
        / card_users (and card_user_counts) links.
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Delete memory node
      tags:
      - memory-nodes
    get:
      description: Returns a memory node by its ID if the authenticated user owns
        it or has an explicit shared grant
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MemoryNodeFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get memory node by ID
      tags:
      - memory-nodes
  /memory-node/{id}/cards:
    post:
      consumes:
      - application/json
      description: Creates multiple text-only cards under a memory node. Body may
        be a cards array or an object with cards. Each card requires questionTextItems
        and answerTextItems. Pass ?shared=true to mark all cards as shared.
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      - description: Mark all cards as shared
        in: query
        name: shared
        type: boolean
      - description: Cards array or wrapper
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BulkTextCardsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Bulk create text cards
      tags:
      - cards
  /memory-node/{id}/parents-path:
    get:
      description: Returns the first-parent chain from root to the given node (inclusive)
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MemoryNodePathItem'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get parents path for a memory node
      tags:
      - memory-nodes
  /memory-node/{id}/users:
    post:
      consumes:
      - application/json
      description: Creates a memory_node_users row (and card_users for shared cards
        on the node) so the user can see the shared node. Does not make the node visible
        to everyone.
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      - description: User to grant
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GrantMemoryNodeAccessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Grant shared memory node access
      tags:
      - memory-nodes
  /memory-nodes:
    get:
      description: Returns all memory nodes for the authenticated user, or a subset
        when ids query is provided as a JSON array
      parameters:
      - description: JSON array of IDs, e.g. [1,2,3]
        in: query
        name: ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MemoryNodeFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: List memory nodes
      tags:
      - memory-nodes
  /memory-nodes/roots:
    get:
      description: Returns memory nodes with no parents for the authenticated user
        (owned or explicitly granted shared roots)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MemoryNodeFull'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: List root memory nodes
      tags:
      - memory-nodes
  /new-card:
    post:
      consumes:
      - application/json
      description: Creates card items and a card under a memory node for the authenticated
        user
      parameters:
      - description: New card request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.NewCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Create card
      tags:
      - cards
  /new-card-item:
    post:
      consumes:
      - application/json
      description: Creates a new card item owned by the authenticated user
      parameters:
      - description: Card item
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CardItemShort'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardItemFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Create card item
      tags:
      - card-items
...
  /update-cards-field:
    post:
      consumes:
      - application/json
      description: Bulk-updates count on multiple cards owned by the authenticated
        user
      parameters:
      - description: Bulk field update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCardsFieldRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Bulk update card field
      tags:
      - cards
...
schemes:
- http
- https
security:
- Login:
  - api
securityDefinitions:
  Login:
    flow: password
    scopes:
      api: Access the MemoryGuard API
    tokenUrl: /users/token
    type: oauth2
swagger: "2.0"